we are comprehensive
i dont think not explicitly implementing laps invalidates the claim
that this is comprehensive (taking laps is pretty much calling
`elapsed` and collecting the durations, maybe comparing adjacent
items*). this lib offers a fairly wide range of functionality, and
lets u specify exactly how stopwatches are started, stopped, changed,
etc.

* i may add this in the future but right now i dont feel like this is
useful. if ur reading this, and it is, pls reach out or file an issue,
i would be down to implement this if its wanted.